Fix a couple more signed/unsigned compares.
Different compilers find different problems.

+++ crypto/pkcs8/p5_pbev2.c
@@ -356,7 +356,7 @@ static int PKCS5_v2_PBKDF2_keyivgen(EVP_CIPHER_CTX *ctx,
     goto err;
   }
   long iterations = ASN1_INTEGER_get(pbkdf2param->iter);
-  if (iterations <= 0 || iterations > UINT_MAX) {
+  if (iterations <= 0 || iterations > (long) UINT_MAX) {
     OPENSSL_PUT_ERROR(PKCS8, PKCS8_R_BAD_ITERATION_COUNT);
     goto err;
   }
